requirements_cove.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements_cove.in
#
-e file:./
    # via -r requirements_cove.in
asgiref==3.8.1
    # via django
attrs==25.3.0
    # via jsonschema
backports-datetime-fromisoformat==2.0.3
    # via flattentool
bleach==6.2.0
    # via -r requirements_cove.in
btrees==6.1
    # via zodb
cached-property==2.0.1
    # via -r requirements_cove.in
certifi==2025.1.31
    # via
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via persistent
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via rq
commonmark==0.9.1
    # via -r requirements_cove.in
dealer==2.1.0
    # via -r requirements_cove.in
defusedxml==0.7.1
    # via odfpy
django==4.2.20
    # via
    #   -r requirements_cove.in
    #   django-bootstrap3
    #   django-debug-toolbar
    #   django-rq
    #   libcoveweb
django-bootstrap3==25.1
    # via
    #   -r requirements_cove.in
    #   libcoveweb
django-debug-toolbar==5.1.0
    # via -r requirements_cove.in
django-environ==0.12.0
    # via -r requirements_cove.in
django-rq==3.0.0
    # via -r requirements_cove.in
et-xmlfile==2.0.0
    # via openpyxl
flattentool==0.27.0
    # via
    #   -r requirements_cove.in
    #   libcove
    #   libcoveweb
idna==3.10
    # via requests
ijson==3.3.0
    # via
    #   flattentool
    #   lib360dataquality
json-merge-patch==0.3.0
    # via
    #   -r requirements_cove.in
    #   lib360dataquality
jsonref==1.1.0
    # via
    #   flattentool
    #   libcove
jsonschema==3.2.0
    # via
    #   -r requirements_cove.in
    #   lib360dataquality
    #   libcove
libcove==0.31.0
    # via
    #   -r requirements_cove.in
    #   lib360dataquality
    #   libcoveweb
libcoveweb==0.31.1
    # via -r requirements_cove.in
lxml==5.3.1
    # via flattentool
markupsafe==3.0.2
    # via werkzeug
odfpy==1.4.1
    # via flattentool
openpyxl==3.1.5
    # via
    #   -r requirements_cove.in
    #   flattentool
persistent==6.1.1
    # via
    #   btrees
    #   zodb
pycparser==2.22
    # via cffi
pyrsistent==0.20.0
    # via jsonschema
python-dateutil==2.9.0.post0
    # via
    #   -r requirements_cove.in
    #   lib360dataquality
pytz==2025.2
    # via flattentool
rangedict==0.1.7
    # via
    #   -r requirements_cove.in
    #   lib360dataquality
redis==5.2.1
    # via
    #   django-rq
    #   rq
requests==2.32.3
    # via
    #   -r requirements_cove.in
    #   libcove
    #   libcoveweb
rfc3339-validator==0.1.4
    # via libcove
rfc3987==1.3.8
    # via
    #   -r requirements_cove.in
    #   libcove
rq==2.3.1
    # via django-rq
schema==0.7.7
    # via flattentool
sentry-sdk==2.24.1
    # via -r requirements_cove.in
six==1.17.0
    # via
    #   jsonschema
    #   python-dateutil
    #   rfc3339-validator
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
strict-rfc3339==0.7
    # via -r requirements_cove.in
transaction==5.0
    # via zodb
urllib3==2.3.0
    # via
    #   requests
    #   sentry-sdk
webencodings==0.5.1
    # via bleach
werkzeug==3.1.3
    # via libcoveweb
whitenoise==6.9.0
    # via -r requirements_cove.in
xmltodict==0.14.2
    # via
    #   -r requirements_cove.in
    #   flattentool
zc-lockfile==3.0.post1
    # via zodb
zc-zlibstorage==1.2.0
    # via flattentool
zconfig==4.2
    # via zodb
zodb==6.0
    # via
    #   flattentool
    #   zc-zlibstorage
zodbpickle==4.2
    # via zodb
zope-deferredimport==5.0
    # via persistent
zope-interface==7.2
    # via
    #   btrees
    #   persistent
    #   transaction
    #   zc-zlibstorage
    #   zodb
    #   zope-proxy
zope-proxy==6.1
    # via zope-deferredimport

# The following packages are considered to be unsafe in a requirements file:
# setuptools
